requirements.txt
annotated-types==0.7.0
anyio==4.6.0
asyncio==3.4.3
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.4.0
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
grpcio==1.66.2
grpcio-status==1.66.2
h11==0.14.0
httpcore==1.0.6
httplib2==0.22.0
httpx==0.27.2
idna==3.10
packaging==24.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
pyaes==1.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
pyparsing==3.2.0
pytesseract==0.3.13
python-dotenv==1.0.1
python-telegram-bot==21.6
requests==2.32.3
rsa==4.9
setuptools==75.1.0
sniffio==1.3.1
Telethon==1.37.0
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
wheel==0.44.0
